refactor: move ExtensionCapability enum + descriptor to homeboy-extension-contract
Stage-2 cut following the ExtensionManifest move: the ExtensionCapability enum
(Lint/Test/Build/Bench/Fuzz/Trace/Deps), its ExtensionCapabilityDescriptor, and
the full impl ExtensionCapability block all move to the contract crate.

This is now possible because every method on the impl (descriptor, label,
has_manifest_support, script_path, requires_script) only touches
ExtensionManifest - which itself moved to the contract crate. The
manifest-support function pointers (ExtensionManifest::has_lint etc.) resolve
in the leaf crate, so the behavior travels with the data.

Widened the three pub(crate) methods to pub for cross-crate access. Core's
capability.rs re-imports the enum; extension mod.rs re-exports it from the
contract crate, keeping crate::extension::ExtensionCapability stable across all
~17 core consumers.

ExtensionCapability was one of the most-referenced extension symbols (17 core
prod refs) - moving it meaningfully shrinks the extension behavior crate's
eventual reverse-edge surface.
