+++ crates/homeboy-core/src/extension/capability.rs
@@ -7,95 +7,13 @@ use super::manifest::ExtensionManifest;
 use super::registry::{extension_path, load_extension};
 use super::runner::ExtensionRunner;
 use super::ResolvedExtensionInvocationContext;
+use homeboy_extension_contract::ExtensionCapability;
 
 pub(crate) fn stderr_tail(stderr: &str) -> String {
     const MAX_LINES: usize = 20;
     let lines: Vec<&str> = stderr.lines().collect();
     let start = lines.len().saturating_sub(MAX_LINES);
     lines[start..].join("\n")
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ExtensionCapability {
-    Lint,
-    Test,
-    Build,
-    Bench,
-    Fuzz,
-    Trace,
-    Deps,
-}
-
-/// Static metadata for an [`ExtensionCapability`] variant.
-///
-/// Centralizing label, manifest-support probe, and script accessor in one
-/// descriptor keeps variant additions localized: a new capability only
-/// needs one new arm in [`ExtensionCapability::descriptor`] instead of
-/// parallel arms scattered across each getter / policy method.
-struct ExtensionCapabilityDescriptor {
-    label: &'static str,
-    has_manifest_support: fn(&ExtensionManifest) -> bool,
-    script_path: fn(&ExtensionManifest) -> Option<&str>,
-}
-
-impl ExtensionCapability {
-    fn descriptor(self) -> ExtensionCapabilityDescriptor {
-        match self {
-            ExtensionCapability::Lint => ExtensionCapabilityDescriptor {
-                label: "lint",
-                has_manifest_support: ExtensionManifest::has_lint,
-                script_path: ExtensionManifest::lint_script,
-            },
-            ExtensionCapability::Test => ExtensionCapabilityDescriptor {
-                label: "test",
-                has_manifest_support: ExtensionManifest::has_test,
-                script_path: ExtensionManifest::test_script,
-            },
-            ExtensionCapability::Build => ExtensionCapabilityDescriptor {
-                label: "build",
-                has_manifest_support: ExtensionManifest::has_build,
-                script_path: ExtensionManifest::build_script,
-            },
-            ExtensionCapability::Bench => ExtensionCapabilityDescriptor {
-                label: "bench",
-                has_manifest_support: ExtensionManifest::has_bench,
-                script_path: ExtensionManifest::bench_script,
-            },
-            ExtensionCapability::Fuzz => ExtensionCapabilityDescriptor {
-                label: "fuzz",
-                has_manifest_support: ExtensionManifest::has_fuzz,
-                script_path: ExtensionManifest::fuzz_script,
-            },
-            ExtensionCapability::Trace => ExtensionCapabilityDescriptor {
-                label: "trace",
-                has_manifest_support: ExtensionManifest::has_trace,
-                script_path: ExtensionManifest::trace_script,
-            },
-            ExtensionCapability::Deps => ExtensionCapabilityDescriptor {
-                label: "deps",
-                has_manifest_support: ExtensionManifest::has_deps,
-                script_path: ExtensionManifest::deps_script,
-            },
-        }
-    }
-
-    pub fn label(self) -> &'static str {
-        self.descriptor().label
-    }
-
-    pub(crate) fn has_manifest_support(self, manifest: &ExtensionManifest) -> bool {
-        (self.descriptor().has_manifest_support)(manifest)
-    }
-
-    pub(crate) fn script_path(self, manifest: &ExtensionManifest) -> Option<&str> {
-        (self.descriptor().script_path)(manifest)
-    }
-
-    pub(crate) fn requires_script(self) -> bool {
-        // Fuzz supports manifest-only workload discovery; `fuzz run` validates
-        // its runner script before execution.
-        !matches!(self, ExtensionCapability::Build | ExtensionCapability::Fuzz)
-    }
 }
 
 #[derive(Debug, Clone)]
